qcacld-3.0: Do not strip RSN IE in RSO config IEs
During OWE-MLO roaming, the vdev-1 will be brought up after
the key-exchange completion on the assoc vdev. The connect
request for the vdev-1 bring up is generated from the RSO
config of the assoc vdev. In current design, the assoc IEs
stored in the RSO config do not have RSN IE. Therefore, the
vdev-1 bring up would not be populated with certain crypto
params generated out of RSN IE. Eg: mgmt cipher.

To fix this, always cache the unstripped version of the assoc
IEs from userspace in the RSO CFG.

Change-Id: I76fec19a454f4b145bea7f2bf2d9f3b45e9400b6
CRs-Fixed: 3821233